Use Signed 32 bit integer type for ids
Postgres does not support unsigned numbers (besides the oid type).
As such, all ids in IML are actually signed 32 bit ints.

Update the representation of these types in wire-types and other spots to reflect this.

This will pave the way for using diesel struct types instead of
repeating types in both iml-orm and wire-types.

+++ iml-api/src/command.rs
@@ -31,7 +31,7 @@ pub(crate) async fn get_command(pool: &DbPool, id: i32) -> Result<Command, ImlAp
         complete: cmd.complete,
         created_at: format!("{}", cmd.created_at),
         errored: cmd.errored,
-        id: cmd.id as u32,
+        id: cmd.id,
         message: cmd.message,
         resource_uri: format!("/api/{}/{}/", Command::endpoint_name(), cmd.id),
         jobs: jobs
